优化mount操作

+++ drivers/gps/lidbg_gps.c
@@ -392,17 +392,25 @@ static int  gps_probe(struct platform_device *pdev)
 #ifndef SOC_msm8x25
     if(is_ublox_exist() < 0)
     {
-		lidbg_shell_cmd("mount -o remount /system");
-		lidbg_shell_cmd("rm /flysystem/lib/out/"FLY_GPS_SO);
-		lidbg_shell_cmd("rm /system/lib/modules/out/"FLY_GPS_SO);
-		lidbg_shell_cmd("mv /flysystem/lib/hw/"FLY_GPS_SO"  /flysystem/lib/hw/gps.soc.bak");
+	    if(g_var.is_first_update)
+	    {
+			lidbg_shell_cmd("mount -o remount /system");
+			lidbg_shell_cmd("mount -o remount /flysystem");
+			
+			lidbg_shell_cmd("rm /flysystem/lib/out/"FLY_GPS_SO);
+			lidbg_shell_cmd("rm /system/lib/modules/out/"FLY_GPS_SO);
+			lidbg_shell_cmd("mv /flysystem/lib/hw/"FLY_GPS_SO"  /flysystem/lib/hw/gps.soc.bak");
+	    }
 		lidbg("[ublox]ublox.miss\n\n");
 		return 0;
     }
     else
     {
-
-		lidbg_shell_cmd("mv /flysystem/lib/hw/gps.soc.bak  /flysystem/lib/hw/"FLY_GPS_SO);
+	    if(g_var.is_first_update)
+	    {
+	    	lidbg_shell_cmd("mount -o remount /flysystem");
+			lidbg_shell_cmd("mv /flysystem/lib/hw/gps.soc.bak  /flysystem/lib/hw/"FLY_GPS_SO);
+	    }
         lidbg("[ublox]ublox.exist\n\n");
         fs_mem_log("ublox_exist=true\n");
